requirements.txt
numpy
pandas
Flask
tensorflow-cpu
pickle
requests
Pillow
gunicorn
opencv-python